service_test: Add
