+++ pkg/controller/controller_test.go
@@ -17,7 +17,6 @@
 package controller
 
 import (
-	"fmt"
 	"testing"
 	"time"
 
@@ -26,7 +25,6 @@ import (
 	kubeinformers "k8s.io/client-go/informers"
 	kubeclientset "k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/cache"
 
 	tfv1alpha2 "github.com/kubeflow/tf-operator/pkg/apis/tensorflow/v1alpha2"
 	tfjobclientset "github.com/kubeflow/tf-operator/pkg/client/clientset/versioned"
@@ -100,90 +98,12 @@ func newTFJob(worker, ps int) *tfv1alpha2.TFJob {
 }
 
 func getKey(tfJob *tfv1alpha2.TFJob, t *testing.T) string {
-	if key, err := KeyFunc(tfJob); err != nil {
+	key, err := KeyFunc(tfJob)
+	if err != nil {
 		t.Errorf("Unexpected error getting key for job %v: %v", tfJob.Name, err)
 		return ""
-	} else {
-		return key
 	}
-}
-
-func newBasePod(name string, tfJob *tfv1alpha2.TFJob, t *testing.T) *v1.Pod {
-	return &v1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:            name,
-			Labels:          genLabels(getKey(tfJob, t)),
-			Namespace:       tfJob.Namespace,
-			OwnerReferences: []metav1.OwnerReference{*metav1.NewControllerRef(tfJob, controllerKind)},
-		},
-	}
-}
-
-func newPod(tfJob *tfv1alpha2.TFJob, typ string, index int, t *testing.T) *v1.Pod {
-	pod := newBasePod(fmt.Sprintf("%s-%d", typ, index), tfJob, t)
-	pod.Labels[tfReplicaTypeLabel] = typ
-	pod.Labels[tfReplicaIndexLabel] = fmt.Sprintf("%d", index)
-	return pod
-}
-
-// create count pods with the given phase for the given tfJob
-func newPodList(count int32, status v1.PodPhase, tfJob *tfv1alpha2.TFJob, typ string, start int32, t *testing.T) []*v1.Pod {
-	pods := []*v1.Pod{}
-	for i := int32(0); i < count; i++ {
-		newPod := newPod(tfJob, typ, int(start+i), t)
-		newPod.Status = v1.PodStatus{Phase: status}
-		pods = append(pods, newPod)
-	}
-	return pods
-}
-
-func setPodsStatuses(podIndexer cache.Indexer, tfJob *tfv1alpha2.TFJob, typ string, pendingPods, activePods, succeededPods, failedPods int32, t *testing.T) {
-	var index int32
-	for _, pod := range newPodList(pendingPods, v1.PodPending, tfJob, typ, index, t) {
-		podIndexer.Add(pod)
-	}
-	index += pendingPods
-	for _, pod := range newPodList(activePods, v1.PodRunning, tfJob, typ, index, t) {
-		podIndexer.Add(pod)
-	}
-	index += activePods
-	for _, pod := range newPodList(succeededPods, v1.PodSucceeded, tfJob, typ, index, t) {
-		podIndexer.Add(pod)
-	}
-	index += succeededPods
-	for _, pod := range newPodList(failedPods, v1.PodFailed, tfJob, typ, index, t) {
-		podIndexer.Add(pod)
-	}
-}
-
-func newService(tfJob *tfv1alpha2.TFJob, typ string, index int, t *testing.T) *v1.Service {
-	service := &v1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:            fmt.Sprintf("%s-%d", typ, index),
-			Labels:          genLabels(getKey(tfJob, t)),
-			Namespace:       tfJob.Namespace,
-			OwnerReferences: []metav1.OwnerReference{*metav1.NewControllerRef(tfJob, controllerKind)},
-		},
-	}
-	service.Labels[tfReplicaTypeLabel] = typ
-	service.Labels[tfReplicaIndexLabel] = fmt.Sprintf("%d", index)
-	return service
-}
-
-// create count pods with the given phase for the given tfJob
-func newServiceList(count int32, tfJob *tfv1alpha2.TFJob, typ string, t *testing.T) []*v1.Service {
-	services := []*v1.Service{}
-	for i := int32(0); i < count; i++ {
-		newService := newService(tfJob, typ, int(i), t)
-		services = append(services, newService)
-	}
-	return services
-}
-
-func setServices(serviceIndexer cache.Indexer, tfJob *tfv1alpha2.TFJob, typ string, activeWorkerServices int32, t *testing.T) {
-	for _, service := range newServiceList(activeWorkerServices, tfJob, typ, t) {
-		serviceIndexer.Add(service)
-	}
+	return key
 }
 
 func getCondition(tfJob *tfv1alpha2.TFJob, condition tfv1alpha2.TFJobConditionType, reason string) bool {
